refactors adding comments to posts
